backend/requirements.txt
asgiref==3.10.0
awsebcli==3.25.2
bcrypt==5.0.0
blessed==1.22.0
botocore==1.40.49
cement==2.10.14
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.2
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fabric==3.2.2
Faker==37.12.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
invoke==2.2.0
jmespath==1.0.1
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pillow==11.3.0
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2-binary==2.9.11
pycparser==2.23
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.14
whitenoise==6.11.0
wrapt==1.17.3
